Fix code for some compiler warnings
- Inside the Sd2Card::setSckRate() function implementaion, a variable
'int v' is declared before a #ifdef directive, and assigned inside
the directive statemnt. This caused the compiler to issue a warning
that the varibale might be left uninitialized. The variable is now
initilized to 0.

- Inside the Sd2Card::readData() function implementaion, a variable
'uint16_t n' is declared before a #ifdef directive, and assigned
and used inside the directive statemnt. This caused the compiler
to issue a warning that the variable might be left unused. The
variable is now declared and asigned inside the #idfef directive.

- Inside the SdFile::rmRfStar() function implementation, the following
comparison is made 'curPosition_ != (32*(index + 1))'. curPosition_ is
of type unit32_t but the using of int constants in the right side
of the comparison causes a compiler warning that signed and unsigned
integer types are being compared. The integer constants are now written
as unsigned.